refactor(models): move model files into contracts directory
