+++ src/contracts/model.spec.ts
