+++ src/contracts/model.ts
